added authentication and authorization

+++ routes/listings.js
@@ -4,6 +4,8 @@ const Listing = require("../models/listing")
 const ExpressError = require("../utils/error.js")
 const wrapAsync = require("../utils/wrapAsync.js")
 const {listingSchema, reviewSchema} = require("../schema.js")
+const flash = require("connect-flash")
+const {checkAuth, getUrl} = require("../middleware.js")
 
 
 
@@ -28,6 +30,7 @@ router.get
     "/",
     wrapAsync(async (req, res)=>
     {
+        // console.log(req.user)
         const result = await Listing.find({})
         res.render("listing.ejs", {result})
     })
@@ -38,9 +41,12 @@ router.get
 router.get
 (
     "/new",
-    (req, res)=>
+    checkAuth,
+    async (req, res)=>
     {
+        
         res.render("new.ejs")
+
     }
 )
 
@@ -49,11 +55,15 @@ router.get
 router.post
 (
     "/", 
+    checkAuth,
     validateListing,
     wrapAsync(async (req, res, next)=>
     {
         const item = new Listing(req.body)
+        item.owner = req.user
         await item.save()
+
+        req.flash("success", "Successfully added a new listing!")
         res.redirect("/listings")
     })
 )
@@ -65,8 +75,29 @@ router.get
     "/:id", 
     wrapAsync(async (req, res)=>
     {
-        const item = await Listing.findById(req.params.id).populate("reviews")
-        res.render("show.ejs", {item})
+        const item = await Listing.findById(req.params.id)
+                                            .populate
+                                            ({
+                                                path: "reviews",
+                                                populate: {path: "user"},
+                                            })
+                                            .populate("owner")
+        
+
+        if (!item)
+        {
+            req.flash("error", "Listing not found!")
+            res.redirect("/listings")
+        }
+
+        else
+        {
+            // console.log(req.user._id)
+            // console.log(item.owner._id)
+            res.render("show.ejs", {item})
+        }
+
+            
     })
 )
 
@@ -76,11 +107,25 @@ router.get
 router.get
 (
     "/:id/edit", 
+    checkAuth,
     wrapAsync(async (req, res)=>
     {
-        const item = await Listing.findById(req.params.id)
-        res.render("edit.ejs", {item})
-        // res.send("hi")
+        const item = await Listing.findById(req.params.id).populate("owner")
+
+        if (!item)
+        {
+            req.flash("error", "Listing not found!")
+            res.redirect("/listings")
+        }
+
+        else if (req.user._id.equals(item.owner._id))
+            res.render("edit.ejs", {item})
+
+        else
+        {
+            req.flash("error", "You don't have access to edit this!")
+            res.redirect(`/listings/${req.params.id}`)
+        }
     })
 )
 
@@ -89,10 +134,20 @@ router.get
 router.patch
 (
     "/:id", 
+    checkAuth,
     validateListing,
     wrapAsync(async (req, res, next)=>
     {
-        const item = await Listing.findByIdAndUpdate(req.params.id, req.body, {new: true, runValidators: true})
+        const item = await Listing.findById(req.params.id).populate("owner")
+
+        if (!req.user._id.equals(item.owner._id))
+        {
+            req.flash("error", "You don't have access to edit this!")
+            res.redirect(`/listings/${req.params.id}`)
+        }
+
+        await Listing.findByIdAndUpdate(req.params.id, req.body, {new: true, runValidators: true})
+        req.flash("success", "Successfully updated the listing!")
         res.redirect(`/listings/${item._id}`)
     })
 )
@@ -102,11 +157,31 @@ router.patch
 router.delete
 (
     "/:id", 
+    checkAuth,
     wrapAsync(async (req, res, next)=>
     {
-        const item = await Listing.findByIdAndDelete(req.params.id)
-        res.redirect("/listings")
-        // res.send("working")
+        const item = await Listing.findById(req.params.id).populate("owner")
+        
+        if (!item)
+        {
+            req.flash("error", "Listing not found!")
+            res.redirect("/listings")
+        }
+        
+        else if (req.user._id.equals(item.owner._id))
+        {
+            await Listing.findByIdAndDelete(req.params.id)
+            req.flash("success", "Successfully deleted a listing!")
+            res.redirect("/listings")
+        }
+
+        else
+        {
+            req.flash("error", "You don't have access to delete this!")
+            res.redirect(`/listings/${req.params.id}`)
+        }
+        
+        
     })
 )
