+++ routes/review.js
@@ -5,6 +5,7 @@ const ExpressError = require("../utils/error.js")
 const Review = require("../models/review.js")
 const Listing = require("../models/listing")
 const {reviewSchema, listingSchema} = require("../schema.js")
+const { checkAuth } = require("../middleware.js")
 
 
 
@@ -27,17 +28,22 @@ function validateReview (req, res, next)
 router.post
 (
     "/",
+    checkAuth,
     validateReview,
     wrapAsync (async (req, res)=>
     {
         const item = await Listing.findById(req.params.id)
         const review = new Review(req.body)
+        review.user = req.user._id
+        
         item.reviews.push(review)
 
         await review.save()
         await item.save()
 
-        console.log("<<< review added >>>")
+        req.flash("success", "Successfully added a new review!")
+
+        // console.log("<<< review added >>>")
         res.redirect(`/listings/${req.params.id}`)
     })
 )
@@ -47,13 +53,31 @@ router.post
 router.delete
 (
     "/:reviewId",
+    checkAuth,
     wrapAsync(async (req, res)=>
     {
-        await Listing.findByIdAndUpdate (req.params.id, {$pull: {reviews: req.params.reviewId}})
-        const result = await Review.findByIdAndDelete(req.params.reviewId)
-        console.log("<<< review deleted >>>")
+        const review = await Review.findById(req.params.reviewId).populate("user")
+        if (review && !review.user._id.equals(req.user._id))
+        {
+            req.flash("error", "you don't have access to delete it")
+            res.redirect(`/listings/${req.params.id}`)
+        }
 
-        res.redirect( `/listings/${req.params.id}` )
+        else
+        {const result = await Listing.findByIdAndUpdate (req.params.id, {$pull: {reviews: req.params.reviewId}})
+
+        if (!result)
+        {
+            req.flash("error", "Review not found!")
+            res.redirect(`/listings/${req.params.id}`)
+        }
+
+        else
+        {
+            await Review.findByIdAndDelete(req.params.reviewId)
+            req.flash("success", "Successfully deleted a review!")
+            res.redirect( `/listings/${req.params.id}` )
+        }}
 
     })
 )
